requirements.txt
absl-py==2.3.1
aiobotocore==2.25.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.17.1
altair==5.5.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.4.0
awscli==1.42.71
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.14.2
billiard==4.2.2
binaryornot==0.4.4
bleach==6.3.0
blinker==1.9.0
blis==1.3.0
boto3==1.40.70
botocore==1.40.71
branca==0.8.2
cachetools==6.2.1
catalogue==2.0.10
catboost==1.2.8
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
configobj==5.0.9
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
dacite==1.6.0
dagshub==0.6.3
dagshub-annotation-converter==0.1.15
databricks-sdk==0.73.0
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docutils==0.19
dpath==2.2.0
dulwich==0.24.10
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
emoji==2.15.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
executing==2.2.1
fastapi==0.121.1
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
flatten-dict==0.4.2
flufl.lock==8.2.0
folium==0.20.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
funcy==2.0
fuzzywuzzy==0.18.0
gast==0.6.0
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-pasta==0.2.0
gql==4.0.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.75.1
gto==1.9.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
huey==2.5.4
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.11
ImageHash==4.3.1
importlib_metadata==8.7.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.15
keras==3.11.3
kiwisolver==1.4.9
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
lark==1.3.1
Levenshtein==0.27.1
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.44.0
lxml==6.0.2
Mako==1.3.10
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
minify_html==0.16.4
mistune==3.1.4
ml_dtypes==0.5.3
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
mpmath==1.3.0
multidict==6.7.0
multimethod==1.12
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
notebook==7.4.7
notebook_shim==0.2.4
numba==0.61.0
numpy==2.1.1
omegaconf==2.3.0
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.4
osmnx==2.0.6
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pathvalidate==3.3.1
patsy==1.0.1
phik==0.12.5
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.12.0
pydantic_core==2.41.4
pydeck==0.9.1
pydot==4.0.1
pygit2==1.19.0
Pygments==2.19.2
pygtrie==2.5.0
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-Levenshtein==0.27.1
python-slugify==8.0.4
pytz==2025.2
PyWavelets==1.9.0
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.7.2
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
s3fs==2025.9.0
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.5.2
scipy==1.15.3
scmrepo==3.5.2
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
sentence-transformers==5.1.1
setuptools==80.9.0
shap==0.49.1
shapely==2.1.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
slicer==0.0.8
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlparse==0.5.3
sqltrie==0.11.2
srsly==2.5.1
stack-data==0.6.3
starlette==0.49.3
statsmodels==0.14.5
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
treelib==1.8.0
typeguard==4.4.4
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
visions==0.8.1
voluptuous==0.15.2
waitress==3.0.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.1
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.3
xgboost==2.1.1
xlsxwriter==3.2.9
xyzservices==2025.4.0
yarl==1.22.0
ydata-profiling==4.17.0
zc.lockfile==4.0
zipp==3.23.0
